requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
pydantic==1.10.12
python-dotenv==1.0.0
google-generativeai==0.3.1
aiohttp==3.8.6
sqlalchemy==2.0.22
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.0
python-multipart==0.0.6
langchain_core
langgraph
langchain_community
langchain_anthropic
langsmith